Handles (almost) all the error codes and removes initialization checks
Attempting to use the pump prior initalization triggers the error "g",
therefore after the introduction of PumpHWError, ControllerInitError is
no more needed.

This allows to simplify the code in an EAFP way and gives more
diagnostic error messages. (Pump number seems to be missing though!)

Intended behaviour checked on the device.