Introduces parallel transfer
Unfortunately, if a dict of pumps and volumes only are passed, the valve
in and out have to be the same for all the pumps.

This is usually not a big deal as parallel transfer is performed among
group of similar pumps (oils, stock solutions etc) and it can be an acceptable
limitation.

Note that the previous version of transfer was already compatible with
parallel transfers if a list of pumps was passed as parameter, however
the volume across the different pumps had to be the same. This removes
only such limitation.

Friday afternoon considerations:
All of this is needed due to the blocking nature of transfer that has a
wait between pump and deliver.
If the C3000Controller was re-designed with threading so that each pump
is a thread that get locked during long transfer operation with only
terminate commands allowed that would fix all the problems.
A major re-design of pycont would be needed though.